# Conflicts:
#   .gitignore
#   kernel/include/syscall/syscall_defs.hh
#   user/user_init.c

+++ kernel/pm/process_manager.cc
@@ -1731,10 +1731,6 @@ namespace pm
 		else // normal file
 		{
 			fs::normal_file *f = new fs::normal_file( attrs, dentry );
-			if ( dentry->rName() == "test.txt" ) {
-				// printf( "dup success\n" );
-				dentry->dup();
-			}
 			// log_info( "test normal file read" );
 			// {
 			// 	fs::file *ff = ( fs::file * ) f;
@@ -2026,6 +2022,7 @@ namespace pm
 				return -1;
 
 			if ( path[0] == '.' && path[1] == '/' ) path = path.substr( 2 );
+			if ( path == "/proc/interrupts" ) return -1;
 
 			return fs::k_file_table.unlink( path );
 		}
